refactor(market_research): CQRS 패턴 -> 통합
- 처음에는 확장성을 생각하여 CQRS 패터을 적용하였으나 규모가 아직 크지않고 이슈가 생기는 경우 수정하는 것이 낫다고 판단되어 통합하는 형태로 수정했습니다.
