+++ src/test/java/org/jiwoo/back/business/service/BusinessServiceImplTest.java
@@ -1,8 +1,8 @@
-package org.jiwoo.back.business.query.service;
+package org.jiwoo.back.business.service;
 
-import org.jiwoo.back.business.query.aggregate.entity.Business;
-import org.jiwoo.back.business.query.dto.BusinessDTO;
-import org.jiwoo.back.business.query.repository.BusinessMapper;
+import org.jiwoo.back.business.aggregate.entity.Business;
+import org.jiwoo.back.business.dto.BusinessDTO;
+import org.jiwoo.back.business.repository.BusinessMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
